requirements.txt
pandas==1.5.3
requests==2.28.2
numpy==1.24.2
pyyaml==6.0
python-dotenv==1.0.0
matplotlib==3.7.1
apscheduler==3.11.0